Crate octets

Structs

BufferTooShortError

An error indicating that the provided OctetsMut is not big enough.

Octets

A zero-copy immutable byte buffer.

OctetsMut

A zero-copy mutable byte buffer.

Constants

MAX_VAR_INT

Maximum value that can be encoded via varint.

Functions

huffman_encoding_len

Returns how long the Huffman encoding of the given buffer will be.

varint_len

Returns how many bytes it would take to encode v as a variable-length integer.

varint_parse_len

Returns how long the variable-length integer is, given its first byte.

Type Aliases

Result

A specialized Result type for OctetsMut operations.